docs: 修改readme

+++ index.js
@@ -7,6 +7,6 @@ module.exports = {
   rules: {
     'no-console': 'off',
     'indent': [ 'error', 2 ],
-    'quotes': [ 'error', 'single' ],
+    'quotes': [ 'error', 'double' ],
   },
 }
